cleaned up tests file, added separate file for db connect for further testing
