+++ data/seed_events.json
@@ -1,7 +1,7 @@
 {
-    "name": "Rusty2",
-    "date": "Dog",
-    "generalLocation": "6 years",
-    "specificLocation": "a very friendly dog",
+    "name": "Event 1",
+    "date": "September 5",
+    "generalLocation": "NSW",
+    "specificLocation": "12 Street St.",
     "capacity": 10
 }
