+++ test/events.test.js
@@ -1,34 +1,10 @@
 const mongoose = require('mongoose')
 const expect = require("expect")
 const Event = require("../models/event")
-const {
-    getAllEvents, 
-    addEvent, 
-    getEventById, 
-    updateEvent, 
-    deleteEvent} = require("../utils/event_utilities")
+const utilities = require("../utils/event_utilities")
+const {connectToDb} = require("./common/common")
 
-const dbConn = 'mongodb://localhost/secret_gigs_test'
-
-function connectToDb(done) {mongoose.connect(
-    dbConn,
-    {
-        useNewUrlParser : true,
-        useUnifiedTopology : true,
-        useFindAndModify : false
-    }, 
-    err => {
-        if (err){
-            console.log("error connecting database", err)
-            done()
-        } else{
-            console.log("Connected to SG Database!")
-            done()
-        }
-    }
-)}
 let eventId = null
-
 
 before(done => {
     connectToDb(done)
@@ -38,7 +14,6 @@ before(done => {
 	mongoose.disconnect(() => done())
 })
 
-// Setup and tear down functions
 function setupData() {
 	let testEvent = {}
 	testEvent.name = "tester"
@@ -58,20 +33,18 @@ beforeEach( async () => {
 
 })
 
-
 describe("getAllEvents", () => {
     let req = {
             query: {}
         }
 	it("should get all events", async () => {
-        await getAllEvents(req).exec((err,event) =>{
+        await utilities.getAllEvents(req).exec((err,event) =>{
             expect(Object.keys(event).length).toBe(1)
         })
-        // Pass an empty req object
 		
 	})
 	it("name of first event should be tester", async () => {
-        await getAllEvents(req).exec((err,event) => {
+        await utilities.getAllEvents(req).exec((err,event) => {
             expect(event[0].name).toBe("tester")
         })
 	})
@@ -84,7 +57,7 @@ describe("getEventById", () => {
                 id: eventId
             }
         }
-        await getEventById(req).exec((err, event) => {
+        await utilities.getEventById(req).exec((err, event) => {
             expect(event.name).toBe("tester")
         })
     })
@@ -101,7 +74,7 @@ describe("addEvent", () => {
                 capacity: 10
             }
         }
-        await addEvent(req).save((err, event) => {
+        await utilities.addEvent(req).save((err, event) => {
             expect(event.name).toBe(req.body.name)
         })
     })
@@ -114,7 +87,7 @@ describe("deleteEvent", () => {
                 id: eventId
             }
         }
-        await deleteEvent(req).exec();
+        await utilities.deleteEvent(req).exec();
         await Event.findById(req.params.id).exec((err, event) => {
             expect(event).toBe(null);
         })
@@ -134,9 +107,9 @@ describe("updateEvent", () => {
                 specificLocation: "way up high",
                 capacity: 10
             }
-        };
-        await updateEvent(req).exec((err, event) => {
-            expect(event.name).toBe(req.body.name);
-        });
-    });
-});
+        }
+        await utilities.updateEvent(req).exec((err, event) => {
+            expect(event.name).toBe(req.body.name)
+        })
+    })
+})
